#25 - implemented logic for Posts: GetAll, GetById, AddComment, Like, Unlike

+++ TFMovies.API/Services/Implementations/PostService.cs
@@ -1,9 +1,14 @@
-﻿using System.Data;
+﻿using Microsoft.EntityFrameworkCore;
+using System.Data;
+using System.Linq.Expressions;
 using System.Net;
 using System.Security.Claims;
 using TFMovies.API.Common.Constants;
 using TFMovies.API.Data.Entities;
+using TFMovies.API.Data.Entitiesl;
 using TFMovies.API.Exceptions;
+using TFMovies.API.Extensions;
+using TFMovies.API.Filters;
 using TFMovies.API.Models.Requests;
 using TFMovies.API.Models.Responses;
 using TFMovies.API.Repositories.Interfaces;
@@ -18,26 +23,35 @@ public class PostService : IPostService
     private readonly IThemeRepository _themeRepository;
     private readonly ITagRepository _tagRepository;
     private readonly IPostTagRepository _postTagRepository;
+    private readonly IPostCommentRepository _postCommentRepository;
+    private readonly IPostLikeRepository _postLikeRepository;
 
     public PostService(
         IUserRepository userRepository,
         IThemeRepository themeRepository,
         IPostRepository postRepository,
         ITagRepository tagRepository,
-        IPostTagRepository postTagRepository)
+        IPostTagRepository postTagRepository,
+        IPostCommentRepository postCommentRepository,
+        IPostLikeRepository postLikeRepository)
     {
         _userRepository = userRepository;
         _themeRepository = themeRepository;
         _postRepository = postRepository;
         _tagRepository = tagRepository;
         _postTagRepository = postTagRepository;
+        _postCommentRepository = postCommentRepository;
+        _postLikeRepository = postLikeRepository;
     }
 
     public async Task<PostCreateResponse> CreateAsync(PostCreateRequest model, ClaimsPrincipal currentUserPrincipal)
     {
-        var userId = currentUserPrincipal.FindFirstValue("sub");
-    
-        var currentUser = await GetUserByIdAsync(userId);       
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
+
+        if (currentUser == null)
+        {
+            throw new ServiceException(HttpStatusCode.Unauthorized, ErrorMessages.UserNotFound);
+        }
 
         var theme = await GetThemeByNameAsync(model.Theme);
 
@@ -52,15 +66,15 @@ public class PostService : IPostService
             UpdatedAt = DateTime.UtcNow
         };
 
-        
+
         await _postRepository.CreateAsync(post);
 
         var existingTags = await GetOrCreateTagsAsync(model.Tags);
-        
+
         if (existingTags != null)
         {
             await AddPostTagRelationAsync(post, existingTags);
-        }                       
+        }
 
         var response = new PostCreateResponse
         {
@@ -71,12 +85,12 @@ public class PostService : IPostService
             CreatedAt = post.CreatedAt,
             Author = currentUser.Nickname,
             Theme = theme.Name,
-            Tags = existingTags.Select(t => t.Name).ToList()
+            Tags = existingTags?.Select(t => t.Name).ToList() ?? new List<string>()
         };
 
         return response;
     }
-    
+
     public async Task<PostUpdateResponse> UpdateAsync(string id, PostUpdateRequest model)
     {
         var postDb = await GetPostByIdAsync(id);
@@ -92,8 +106,8 @@ public class PostService : IPostService
         postDb.UpdatedAt = DateTime.UtcNow;
 
         var existingTags = await GetOrCreateTagsAsync(model.Tags);
-        
-        await UpdatePostTagRelationAsync(postDb, existingTags);        
+
+        await UpdatePostTagRelationAsync(postDb, existingTags);
 
         await _postRepository.UpdateAsync(postDb);
 
@@ -112,8 +126,192 @@ public class PostService : IPostService
         return response;
     }
 
-    //helpers  
+    public async Task<PostGetAllResponse> GetAllAsync(int page, int limit, string? sort, string? theme, ClaimsPrincipal currentUserPrincipal)
+    {
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
 
+        IQueryable<Post> query = _postRepository.Query()
+            .Include(p => p.User)
+            .Include(p => p.Theme)
+            .Include(p => p.PostTags)
+                .ThenInclude(pt => pt.Tag)
+            .Include(p => p.PostLikes);
+
+        if (!string.IsNullOrEmpty(theme))
+        {
+            var themeDb = await GetThemeByNameAsync(theme);
+            query = query.Where(post => post.ThemeId == themeDb.Id);
+        }
+
+        Expression<Func<Post, object>> sortSelector;
+        switch (sort?.ToLower())
+        {
+            case "rated":
+                sortSelector = p => p.PostLikes.Count;
+                break;
+            default:
+                sortSelector = p => p.CreatedAt; // default
+                break;
+        }
+
+        var pagination = new PaginationFilter
+        {
+            Limit = limit,
+            Page = page
+        };
+
+        var pagedResult = await query.GetPagedDataAsync(pagination, sortSelector, "desc");
+
+        var data = pagedResult.Data.Select(p => new PostShortInfoDto
+        {
+            Id = p.Id,
+            CoverImageUrl = p.CoverImageUrl,
+            Title = p.Title,
+            CreatedAt = p.CreatedAt,
+            Author = p.User.Nickname,
+            IsLiked = (currentUser == null || p.PostLikes == null) ? false : p.PostLikes.Any(pl => pl.UserId == currentUser.Id),
+            Tags = GetTagNamesFromPost(p)
+        }).ToList();
+
+        return new PostGetAllResponse
+        {
+            Page = pagedResult.Page,
+            Limit = pagedResult.Limit,
+            TotalPages = pagedResult.TotalPages,
+            TotalRecords = pagedResult.TotalRecords,
+            Theme = theme,
+            Sort = sort,
+            Data = data
+        };
+    }
+
+    public async Task<PostGetByIdResponse> GetByIdAsync(string id, ClaimsPrincipal currentUserPrincipal)
+    {
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
+
+        var post = await _postRepository.Query()
+            .Include(p => p.User)
+            .Include(p => p.Theme)
+            .Include(p => p.PostTags)
+                .ThenInclude(pt => pt.Tag)
+            .Include(p => p.PostComments)
+                .ThenInclude(pc => pc.User)
+            .Include(p => p.PostLikes)
+            .FirstOrDefaultAsync(p => p.Id == id);
+
+        if (post == null)
+        {
+            throw new ServiceException(HttpStatusCode.NotFound, ErrorMessages.PostNotFound);
+        }
+
+        var otherPostsFromAuthor = await _postRepository.Query()
+            .Where(p => p.UserId == post.UserId && p.Id != id)
+            .Select(p => new AuthorOtherPosts
+            {
+                Id = p.Id,
+                Title = p.Title,
+                CreatedAt = p.CreatedAt,
+                Tags = p.PostTags.Select(pt => pt.Tag.Name).ToList()
+            })
+            .OrderByDescending(p => p.CreatedAt)
+            .Take(DefaultPaginationValues.AuthorOtherPostsLimit)
+            .ToListAsync();
+
+        var commentDetails = post.PostComments?.Select(pc => new CommentDetail
+        {
+            Author = pc.Author,
+            Content = pc.Content,
+            CreatedAt = pc.CreatedAt
+        }).ToList();
+
+        var response = new PostGetByIdResponse
+        {
+            Id = post.Id,
+            CoverImageUrl = post.CoverImageUrl,
+            Title = post.Title,
+            Theme = post.Theme.Name,
+            HtmlContent = post.HtmlContent,
+            CreatedAt = post.CreatedAt,
+            AuthorId = post.UserId,
+            Author = post.User.Nickname,
+            IsLiked = (currentUser == null || post.PostLikes == null) ? false : post.PostLikes.Any(pl => pl.UserId == currentUser.Id),
+            LikesCount = post.PostLikes?.Count ?? 0,
+            CommentsCount = post.PostComments?.Count ?? 0,
+            Tags = GetTagNamesFromPost(post),
+            Comments = commentDetails,
+            AuthorOtherPosts = otherPostsFromAuthor
+        };
+
+        return response;
+    }
+    public async Task<PostAddCommentResponse> AddCommentAsync(PostAddCommentRequest model, ClaimsPrincipal currentUserPrincipal)
+    {
+        var postDb = await GetPostByIdAsync(model.PostId);
+
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
+
+        if (currentUser == null)
+        {
+            throw new ServiceException(HttpStatusCode.Unauthorized, ErrorMessages.UserNotFound);
+        }
+
+        var newComment = new PostComment
+        {
+            PostId = model.PostId,
+            UserId = currentUser.Id,
+            Content = model.Content,
+            CreatedAt = DateTime.UtcNow
+        };
+
+        await _postCommentRepository.CreateAsync(newComment);
+
+        return new PostAddCommentResponse
+        {
+            PostId = newComment.PostId,
+            Content = newComment.Content,
+            CreatedAt = newComment.CreatedAt,
+            Author = newComment.Author
+        };
+    }
+
+    public async Task AddLikeAsync(string postId, ClaimsPrincipal currentUserPrincipal)
+    {
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
+
+        if (currentUser == null)
+        {
+            throw new ServiceException(HttpStatusCode.Unauthorized, ErrorMessages.UserNotFound);
+        }
+
+        var newPostLike = new PostLike
+        {
+            PostId = postId,
+            UserId = currentUser.Id
+        };
+
+        await _postLikeRepository.CreateAsync(newPostLike);
+    }
+
+    public async Task RemoveLikeAsync(string postId, ClaimsPrincipal currentUserPrincipal)
+    {
+        var currentUser = await GetUserByIdFromClaimAsync(currentUserPrincipal);
+
+        if (currentUser == null)
+        {
+            throw new ServiceException(HttpStatusCode.Unauthorized, ErrorMessages.UserNotFound);
+        }
+
+        var postLikeDb = await _postLikeRepository.Query()
+                             .Where(pl => pl.PostId == postId && pl.UserId == currentUser.Id)
+                             .FirstOrDefaultAsync();
+
+        if (postLikeDb != null)
+        {
+            await _postLikeRepository.DeleteAsync(postLikeDb);
+        }
+    }
+
+    //helpers 
     private async Task<List<Tag>> GetOrCreateTagsAsync(List<string> tagNames)
     {
         CleanAndDistinctTags(tagNames);
@@ -142,7 +340,7 @@ public class PostService : IPostService
 
     private static List<string> CleanAndDistinctTags(List<string> tagNames)
     {
-        return tagNames?.Where(tn => !string.IsNullOrWhiteSpace(tn)).Distinct().ToList() ?? new List<string>();
+        return tagNames.Where(tn => !string.IsNullOrWhiteSpace(tn)).Distinct().ToList();
     }
 
     private async Task AddPostTagRelationAsync(Post post, List<Tag> tags)
@@ -164,13 +362,13 @@ public class PostService : IPostService
 
         var tagNames = tags.Select(t => t.Name).ToList();
 
-        var tagsToRemove = existingPostTags?.Where(pt => pt.Tag != null && !tagNames.Contains(pt.Tag.Name)).ToList();                            
-       
+        var tagsToRemove = existingPostTags?.Where(pt => pt.Tag != null && !tagNames.Contains(pt.Tag.Name)).ToList();
+
         var tagsToAdd = tags.Where(t => existingPostTags == null || !existingPostTags.Any(pt => pt.TagId == t.Id))
                                     .Select(t => new PostTag { Post = post, Tag = t })
                                     .ToList();
 
-        if (tagsToRemove.Any())
+        if (tagsToRemove != null && tagsToRemove.Any())
         {
             await _postTagRepository.DeleteRangeAsync(tagsToRemove);
         }
@@ -178,8 +376,8 @@ public class PostService : IPostService
         if (tagsToAdd.Any())
         {
             await _postTagRepository.CreateRangeAsync(tagsToAdd);
-        }        
-    }     
+        }
+    }
 
     private async Task<Theme> GetThemeByNameAsync(string themeName)
     {
@@ -202,14 +400,17 @@ public class PostService : IPostService
         return post;
     }
 
-    private async Task<User> GetUserByIdAsync(string id)
+    private async Task<User?> GetUserByIdFromClaimAsync(ClaimsPrincipal currentUserPrincipal)
     {
-        var user = await _userRepository.FindByIdAsync(id);
+        var userId = currentUserPrincipal.FindFirstValue("sub");
 
-        if (user == null)
-        {
-            throw new ServiceException(HttpStatusCode.Unauthorized, ErrorMessages.UserNotFound);
-        }
+        var user = await _userRepository.FindByIdAsync(userId);
+
         return user;
+    }
+
+    private static IEnumerable<string> GetTagNamesFromPost(Post post)
+    {
+        return post.PostTags.Select(pt => pt.Tag.Name).ToList();
     }
 }
